+++ app/Http/Controllers/JobController.php
@@ -5,6 +5,10 @@ namespace App\Http\Controllers;
 use App\Models\Job;
 use Illuminate\Http\Request;
 
+private function checkPerms($perms, $class) {
+    return $this->authorize($perms, $class);
+}
+
 class JobController extends Controller {
     /**
      * Display a listing of the resource.
@@ -12,6 +16,10 @@ class JobController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function index() {
+        if (!checkPerms('viewAny', Job::class)) {
+            abort(403);
+        }
+        
         return Job::all();
     }
 
@@ -22,6 +30,10 @@ class JobController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request) {
+        if (!checkPerms('create', Job::class)) {
+            abort(403);
+        }
+        
         return Job::create($request->all());
     }
 
@@ -32,6 +44,10 @@ class JobController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function show($id) {
+        if (!checkPerms('view', Job::class)) {
+            abort(403);
+        }
+        
         return Job::find($id);
     }
 
@@ -43,6 +59,10 @@ class JobController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id) {
+        if (!checkPerms('update', Job::class)) {
+            abort(403);
+        }
+        
         $job = Job::findOrFail($id); // try to find the Job by id, fail if not found
         $job->update($request->all()); // update the Job with the request data
 
@@ -56,6 +76,10 @@ class JobController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy($id) {
+        if (!checkPerms('delete', Job::class)) {
+            abort(403);
+        }
+        
         $job = Job::findOrFail($id); // try to find the Job by id, fail if not found
         $job->delete(); // delete the Job
