+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,10 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use Illuminate\Http\Request;
 
+private function checkPerms($perms, $class) {
+    return $this->authorize($perms, $class);
+}
+
 class UserController extends Controller {
     /**
      * Display a listing of the resource.
@@ -12,6 +16,10 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function index() {
+        if (!checkPerms('viewAny', User::class)) {
+            abort(403);
+        }
+        
         return User::all();
     }
 
@@ -22,6 +30,10 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request) {
+        if (!checkPerms('create', User::class)) {
+            abort(403);
+        }
+        
         return User::create($request->all());
     }
 
@@ -32,6 +44,10 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function show($id) {
+        if (!checkPerms('view', User::class)) {
+            abort(403);
+        }
+        
         return User::find($id);
     }
 
@@ -43,6 +59,10 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id) {
+        if (!checkPerms('update', User::class)) {
+            abort(403);
+        }
+        
         $user = User::findOrFail($id); // try to find the user by id, fail if not found
         $user->update($request->all()); // update the user with the request data
 
@@ -56,6 +76,10 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy($id) {
+        if (!checkPerms('delete', User::class)) {
+            abort(403);
+        }
+        
         $user = User::findOrFail($id); // try to find the user by id, fail if not found
         $user->delete(); // delete the user
